Fixed problem that curly brackets not electriclly indent
The electric-indent-mode was turned off to let C-j perform its desired behavior: new-line and indent. But it turns out that this also disable the auto indent for curly brackets, which is a greater loss.
The solution for now is to change user behavior to use Enter or C-m or M-j for new-line and indent, and C-j will perform the old Enter function that just open line and not indent.